+++ src/widget/WidgetList.h
@@ -7,25 +7,25 @@
  * @brief Widget that allows a user to select a value from a list.
  * Manages a value within a specified list, allowing cycling through values.
  * @tparam T the type of value
- * @tparam V the type of stored value, the type should be fully compatible with `size_t` type, meaning
- * all arithmetic operations, cast, assignment should be supported for type `V`. For example, T = char*, V = Ref<size_t>.
+ * @tparam V the type of stored value, the type should be fully compatible with `uint8_t` type, meaning
+ * all arithmetic operations, cast, assignment should be supported for type `V`. For example, T = char*, V = Ref<uint8_t>.
  *
  * NOTE: Despite the fact that the class has template type <T> the base
- * Widget will have type <size_t>. It reflects that the value of this widget
+ * Widget will have type <uint8_t>. It reflects that the value of this widget
  * is index in the specified list, not value itself. Also be aware that
- * the ItemWidget's callback should have correct type <size_t>, not <T>.
+ * the ItemWidget's callback should have correct type <uint8_t>, not <T>.
  */
-template <typename T, typename V = size_t>
+template <typename T, typename V = uint8_t>
 class WidgetList : public BaseWidgetValue<V> {
   protected:
-    const size_t size;
+    const uint8_t size;
     const bool cycle;
     const T* values;
 
   public:
     WidgetList(
         const T values[],
-        const size_t size,
+        const uint8_t size,
         const V value,
         const char* format,
         const uint8_t cursorOffset,
@@ -66,7 +66,7 @@ class WidgetList : public BaseWidgetValue<V> {
     }
     void updateValue(const __FlashStringHelper* action) {
         BaseWidgetValue<V>::handleChange();
-        LOG(action, (size_t)this->value);
+        LOG(action, (uint8_t)this->value);
     }
     /**
      * @brief Draw the widget into specified buffer.
@@ -76,7 +76,7 @@ class WidgetList : public BaseWidgetValue<V> {
      */
     uint8_t draw(char* buffer, const uint8_t start) override {
         if (start >= ITEM_DRAW_BUFFER_SIZE) return 0;
-        return snprintf(buffer + start, ITEM_DRAW_BUFFER_SIZE - start, this->format, values[(size_t)this->value]);
+        return snprintf(buffer + start, ITEM_DRAW_BUFFER_SIZE - start, this->format, values[(uint8_t)this->value]);
     }
     bool nextValue() {
         if (this->value + 1 < size) {
@@ -115,15 +115,15 @@ class WidgetList : public BaseWidgetValue<V> {
  * @param callback The callback function to call when the value changes (default: nullptr).
  */
 template <typename T>
-inline WidgetList<T, size_t>* WIDGET_LIST(
+inline WidgetList<T, uint8_t>* WIDGET_LIST(
     const T values[],
-    const size_t size,
-    const size_t value,
+    const uint8_t size,
+    const uint8_t value,
     const char* format = "%s",
     const uint8_t cursorOffset = 0,
     const bool cycle = false,
-    void (*callback)(const size_t&) = nullptr) {
-    return new WidgetList<T, size_t>(values, size, value, format, cursorOffset, cycle, callback);
+    void (*callback)(const uint8_t&) = nullptr) {
+    return new WidgetList<T, uint8_t>(values, size, value, format, cursorOffset, cycle, callback);
 }
 
 /**
@@ -137,16 +137,16 @@ inline WidgetList<T, size_t>* WIDGET_LIST(
  * @param format The format of the value (default: "%s").
  * @param cursorOffset The cursor offset (default: 0).
  * @param cycle Whether to cycle through the list (default: false).
- * @param callback The callback function to call when the value changes (default: nullptr), parameter of callback will be `Ref<size_t>`
+ * @param callback The callback function to call when the value changes (default: nullptr), parameter of callback will be `Ref<uint8_t>`
  */
 template <typename T>
-inline WidgetList<T, Ref<size_t>>* WIDGET_LIST_REF(
+inline WidgetList<T, Ref<uint8_t>>* WIDGET_LIST_REF(
     const T values[],
-    const size_t size,
-    size_t& value,
+    const uint8_t size,
+    uint8_t& value,
     const char* format = "%s",
     const uint8_t cursorOffset = 0,
     const bool cycle = false,
-    void (*callback)(const Ref<size_t>&) = nullptr) {
-    return new WidgetList<T, Ref<size_t>>(values, size, Ref<size_t>(value), format, cursorOffset, cycle, callback);
+    void (*callback)(const Ref<uint8_t>&) = nullptr) {
+    return new WidgetList<T, Ref<uint8_t>>(values, size, Ref<uint8_t>(value), format, cursorOffset, cycle, callback);
 }
